// src/queries/model/channelProposalResponseDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * View Nest API
 * API documentation for View Nest application
 * OpenAPI spec version: 1.0
 */
import type { ChannelProposalResponseDtoAplyStts } from './channelProposalResponseDtoAplyStts';
import type { ChannelProposalResponseDtoMdfcnDt } from './channelProposalResponseDtoMdfcnDt';
import type { ChannelProposalResponseDtoMdfrId } from './channelProposalResponseDtoMdfrId';
import type { ChannelProposalResponseDtoStlmId } from './channelProposalResponseDtoStlmId';
import type { ChannelProposalResponseDtoStlmStts } from './channelProposalResponseDtoStlmStts';
import type { UserResponseDto } from './userResponseDto';

export interface ChannelProposalResponseDto {
  /** 신청 내용 */
  aplyCn: string;
  /** 신청 상태 */
  aplyStts: ChannelProposalResponseDtoAplyStts;
  /** 채널명 */
  chnNm: string;
  /** 채널 제안 계획 ID */
  chnPrpslPlnId: number;
  /** 수정일시 */
  mdfcnDt: ChannelProposalResponseDtoMdfcnDt;
  /**
   * 수정자 ID
   * @nullable
   */
  mdfrId: ChannelProposalResponseDtoMdfrId;
  /** 등록일시 */
  regDt: string;
  /** 등록자 ID */
  rgtrId: number;
  /**
   * 정산 ID
   * @nullable
   */
  stlmId: ChannelProposalResponseDtoStlmId;
  /** 정산 상태 */
  stlmStts?: ChannelProposalResponseDtoStlmStts;
  /** 연관 사용자 ID */
  user: UserResponseDto;
}